[Doctrine] Support for disabling ID generators if ID provided in fixture - part 2 (#89)
When we find an entity with a provided ID, we disable the auto ID generator strategy for the whole class. Unfortunately, an entity can have a provided ID but the next one may not and as a result the persisting will fail because Doctrine expected an ID but none where found.

This issue cannot be fixed because the auto ID generation strategy is a configuration done for the whole class and not on an entity basis. So if this happens, we now bail out with an comprehensive message instead of a cryptic one.